+++ src/features/edit-note/api/useUpdateNoteMutation.ts
@@ -1,13 +1,13 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { UpdateNoteRequest } from "@/entities/note/model/types"
-import { updateNote } from "@/features/edit-note/api/updateNote";
+import { patchNote } from "@/entities/note/api/noteApi";
 import { noteKeys } from "@/entities/note/model/queryKeys";
 
 export const useUpdateNote = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (noteData: UpdateNoteRequest) => updateNote(noteData),
+    mutationFn: (noteData: UpdateNoteRequest) => patchNote(noteData),
     onSuccess: (data, variables) => {
       // 새로운 응답 구조에 맞게 캐시 업데이트
       queryClient.setQueryData(
